cek momo

+++ app/Http/Controllers/KemajuanProyekController.php
@@ -1,20 +1,17 @@
 <?php
-
 namespace App\Http\Controllers;
-
 use Illuminate\Http\Request;
 use App\KemajuanProyek;
 use App\Proyek;
 use App\Pelaksanaan;
 use App\Assignment;
 use App\ListPhoto;
+use App\TipePekerjaan;
 use DB;
 use Validator;
 use Illuminate\Support\Facades\Storage;
-
 class KemajuanProyekController extends Controller
 {
-
     /**
      * Create a new controller instance.
      *
@@ -24,26 +21,21 @@ class KemajuanProyekController extends Controller
     {
         $this->middleware('auth');
     }
-
      /**
      * method untuk melihat daftar informasi kemajuan proyek
      * method untuk melihat daftar kemajuan seluruh proyek yang sedang berjalan
      */
-
     public function viewProyek(){
         $idProyek = Assignment::select('assignments.proyek_id')->where('pengguna_id',\Auth::user()->id)->get();
         $listProyek = Proyek::select('proyeks.*')->whereIn('id',$idProyek)->get(); 
         return view('listProyek', compact('listProyek'));
     }
-
     public function detailProyek($id) {
         $proyek = Proyek::find($id);
         $temp = number_format($proyek->projectValue, 2, ',','.');
         $proyek->projectValue = $temp;     
-
         return view('detailProyek', compact('proyek'));
     }
-
     public function viewKemajuan(){
         if(\Auth::user()->role == 2){
             $idProyeks = Proyek::select('proyeks.id')->where('isLPJExist', 0)->where('approvalStatus',2)->get(); //harus diperbarui sesuai dengan pendifinisian baru
@@ -52,19 +44,14 @@ class KemajuanProyekController extends Controller
                 //dd($proyeks);
                 $proyekPrint = [];
                 
-
                 $idPelaksanaan = Pelaksanaan::select('pelaksanaans.id')->whereIn('proyek_id',$idProyeks)->get();
                 if($idPelaksanaan->isEmpty()){
                     return view('viewAll-EmptyKemajuan', compact('proyeks'));
                 }
-
-
-
                 $pelaksanaans = Pelaksanaan::select('pelaksanaans.*','proyeks.*','kemajuan_proyeks.*')
                 ->join('proyeks','proyeks.id','=','pelaksanaans.proyek_id')
                 ->join('kemajuan_proyeks','kemajuan_proyeks.pelaksanaan_id','=','pelaksanaans.id')
                 ->whereIn('proyek_id',$idProyeks)->get();
-
                 //$kemajuans = KemajuanProyek::select('kemajuan_proyeks.*')->whereIn('pelaksanaan_id',$idPelaksanaan)->get();
                 
                 
@@ -111,25 +98,21 @@ class KemajuanProyekController extends Controller
         }
         
     }
-
     public function viewInfo($id){
         $idPelaksanaan = Pelaksanaan::select('pelaksanaans.id')->where('proyek_id',$id)->get();
         $listInformasi = KemajuanProyek::select('kemajuan_proyeks.*')->whereIn('pelaksanaan_id',$idPelaksanaan)->get();
-        $pelaksanaan = Pelaksanaan::where([['proyek_id','=',$id],['approvalStatus','=',0]])->first();
-
-        if($pelaksanaan == null) {
-            DB::table('pelaksanaans')->insert([
-                'approvalStatus' => 0,
-                'createdDate' => now('GMT+7'),
-                'proyek_id' => $id,
-                'created_at' => now('GMT+7'),
-                'updated_at' => now('GMT+7')
-            ]);
-            $pelaksanaan = Pelaksanaan::where([['proyek_id','=',$id],['approvalStatus','=',0]])->first();         
+        $listPekerjaan = DB::table('jenis_pekerjaan')->select('jenis_pekerjaan.name')->where('proyek_id',$id)->get();
+        $lizWork=array($listPekerjaan->count());
+        //dd($listPekerjaan->count());
+        $counter = 0;
+        foreach($listPekerjaan as $pekerjaan) {
+            $lizWork[$counter] = $pekerjaan->name;
+            $counter++;
+            //dd($counter);
         }
-            return view('listInformasi', compact('listInformasi','pelaksanaan'));
+        //dd($lizWork);
+            return view('listInformasi', compact('listInformasi','lizWork'));
     }
-
     public function detailInfo($id) {
         $idPelaksanaan = KemajuanProyek::select('kemajuan_proyeks.pelaksanaan_id')->where('id',$id)->get();
         $idProyek = Pelaksanaan::select('pelaksanaans.proyek_id')->whereIn('id',$idPelaksanaan)->get();
@@ -140,23 +123,24 @@ class KemajuanProyekController extends Controller
         $tanggalInfo = $informasi->reportDate;
         $tanggal = $this->waktu($tanggalInfo);
         $foto = DB::table('listPhoto')->where('kemajuan_id',$id)->get();
-        //dd($foto);
-        // $banyakFoto = DB::table('listPhoto')->where('kemajuan_id',$id)->get()->count();
-        // $listFoto = array();
-        // for($i=0; i<$banyakFoto; i++){
-        //     $listFoto[$i] = $foto[$i]
-        // }
-        // $daftarFoto = ListPhoto::select('listphoto.*')->where('kemajuan_id',$id);
-        //dd($foto);
-
-        return view('detailInformasi', compact('informasi','proyek','tanggal','foto'));
+        $listPekerjaan = DB::table('jenis_pekerjaan')->select('jenis_pekerjaan.name')->where('proyek_id',$idProyek[0]->proyek_id)->get();
+        $lizWork=array($listPekerjaan->count());
+        //dd($listPekerjaan->count());
+        $counter = 0;
+        foreach($listPekerjaan as $pekerjaan) {
+            $lizWork[$counter] = $pekerjaan->name;
+            $counter++;
+        }
+        return view('detailInformasi', compact('informasi','proyek','tanggal','foto','lizWork'));
     }
-
-    public function tambahInformasi($id){
-        $pelaksanaan = Pelaksanaan::find($id);
-        return view('tambahInformasi',compact('pelaksanaan'));
+    public function tambahInformasi(){
+        $idProyek = Assignment::select('assignments.proyek_id')->where('pengguna_id',\Auth::user()->id)->get();
+        $proyekId = $idProyek[0]->proyek_id;
+        //dd($proyekId);
+        $pekerjaan = DB::table('jenis_pekerjaan')->whereIn('proyek_id',$idProyek)->get();
+        //dd($pekerjaan);
+        return view('tambahInformasi',compact('pekerjaan','proyekId'));
     }
-
     public function tambahFoto($id){
         $kemajuan = KemajuanProyek::find($id);
         $idPelaksanaan = $kemajuan->pelaksanaan_id;
@@ -164,12 +148,10 @@ class KemajuanProyekController extends Controller
         //dd($pelaksanaan);
         return view('tambahFoto',compact('kemajuan','pelaksanaan'));
     }
-
     public function simpanFoto($id,Request $request){
         $validator = Validator::make($request->all(),[
             'file' => 'required|image'
         ]);
-
         if ($request->file != null) {
             foreach($request->file as $file) {
                 $uploadedFile = $file;
@@ -188,38 +170,93 @@ class KemajuanProyekController extends Controller
         }
         return redirect()->action('KemajuanProyekController@detailInfo',['id'=>$id]);
     }
-
     /*
     @param \Illuminate\Http\Request
     @return \Illuminate\Http\Response
     */
-    public function simpanInformasi($id, Request $request){
-
-        $idProyek = Pelaksanaan::select('pelaksanaans.proyek_id')->where('id',$id)->get();
-        $data = json_decode($idProyek);
-
+    public function simpanInformasi(Request $request){
+        $idProyek = Assignment::select('assignments.proyek_id')->where('pengguna_id',\Auth::user()->id)->get();
+        $proyekId = $idProyek[0]->proyek_id;
+        //Bulan awal
+        $pelaksanaan = Pelaksanaan::where([['proyek_id','=',$proyekId]])->first();
+        //dd($pelaksanaan);
+        if ($pelaksanaan == null) {
+            $firstDate = $request->reportdate;
+            //dd($firstDate);
+            $firstMonth = date('m', strtotime($firstDate));
+            $firstYear = date('Y', strtotime($firstDate));
+        }
+        else {
+            $sameIdPelaksanaan = Pelaksanaan::where([['proyek_id','=',$proyekId]])->get();
+            $firstDate = DB::table('kemajuan_proyeks')->select('kemajuan_proyeks.reportDate')->whereIn('pelaksanaan_id',$sameIdPelaksanaan)->min('reportDate');
+            //dd($firstDate);
+            $firstMonth = date('m', strtotime($firstDate));
+            $firstYear = date('Y', strtotime($firstDate));
+        }
+        
+        //Bulan dari Form
+        $inputMonth = date('m', strtotime($request->reportdate));
+        $inputYear = date('Y', strtotime($request->reportdate));
+        //Konversi Bulan
+        $yearGap = $inputYear - $firstYear;
+        if ($yearGap == 0) {
+            $adjustedMonth = ($inputMonth - $firstMonth) + 1;    
+        }
+        else if ($yearGap > 0) {
+            if ($inputMonth == $firstMonth) {
+                $adjustedMonth = ($yearGap * 12) + 1;
+            }
+            else if ($inputMonth > $firstMonth) {
+                $adjustedMonth = ($yearGap * 12) + ($inputMonth-$firstMonth) + 1;
+            }
+            else if ($inputMonth < $firstMonth) {
+                $adjustedMonth = ($yearGap * 12) - ($firstMonth-$inputMonth) + 1;
+            }
+        }
+        //dd($adjustedMonth);
+        $pelaksanaan = Pelaksanaan::where([['proyek_id','=',$proyekId],['approvalStatus','=',0],['bulan','=',$adjustedMonth]])->first();
+        //dd($pelaksanaan);
+        //Bikin LAPJUSIK baru
+        if($pelaksanaan == null) {
+            DB::table('pelaksanaans')->insert([
+                'approvalStatus' => 0,
+                'createdDate' => now('GMT+7'),
+                'bulan'=> $adjustedMonth,
+                'proyek_id' => $proyekId,
+                'created_at' => now('GMT+7'),
+                'updated_at' => now('GMT+7')
+            ]);
+            $pelaksanaan = Pelaksanaan::where([['proyek_id','=',$proyekId],['approvalStatus','=',0],['bulan','=',$adjustedMonth]])->first();
+        }
         $validator = Validator::make($request->all(),[
-            'description' => 'required',
+            'tipepekerjaan' => 'required',
             'reportDate' => 'required',
             'tipeKemajuan' => 'required',
             'value' => 'required',
             'pelaksanaan_id' => 'required',
             'file' => 'required|image'
         ]);
- 
         DB::table('kemajuan_proyeks')->insert([
     		'description' => $request->description,
             'reportDate' => $request->reportdate,
             'tipeKemajuan' => $request->tipekemajuan,
             'value' => $request->nilai,
-            'pelaksanaan_id' => $id,
+            'pekerjaan_id' => $request->tipepekerjaan,
+            'pelaksanaan_id' => $pelaksanaan->id,
             'created_at' => now('GMT+7'),
             'updated_at' => now('GMT+7')
         ]);
-
-        $kemajuans = KemajuanProyek::select('kemajuan_proyeks.*')->where('pelaksanaan_id', $id)->get()->last();
+        $realValue = DB::table('kemajuan_proyeks')->select('kemajuan_proyeks.value')->where('pekerjaan_id',$request->tipepekerjaan)->get();
+        //dd(json_decode($realValue));
+        $combinedValue = 0;
+        foreach($realValue as $value) {
+            $combinedValue += $value->value;
+        }
+        DB::table('jenis_pekerjaan')->where('id',$request->tipepekerjaan)->update([
+            'workCurrentValue' => $combinedValue
+        ]);
+        $kemajuans = KemajuanProyek::select('kemajuan_proyeks.*')->where('pelaksanaan_id', $pelaksanaan->id)->get()->last();
         $kemajuan_id = $kemajuans->id;
-
         if ($request->file != null) {
             foreach($request->file as $file) {
                 $uploadedFile = $file;
@@ -236,65 +273,82 @@ class KemajuanProyekController extends Controller
                 ]);
             }
         }
-
         /*if($request->file!= null) {
             $path = $uploadedFile->store('/files');
         }*/
-        return redirect()->action('KemajuanProyekController@viewInfo',['id'=>$data[0]->proyek_id]);
+        return redirect()->action('KemajuanProyekController@viewInfo',['id'=>$proyekId]);
     }
-
     public function editInformasi($id){
         $idPelaksanaan = KemajuanProyek::select('kemajuan_proyeks.pelaksanaan_id')->where('id',$id)->get();
         $idProyek = Pelaksanaan::select('pelaksanaans.proyek_id')->whereIn('id',$idPelaksanaan)->get();
         $proyek = Proyek::find($idProyek[0]->proyek_id);
         $kemajuans = KemajuanProyek::find($id);
-
         $foto = DB::table('listPhoto')->select('listPhoto.*')->where('kemajuan_id',$id)->get();
-        //dd($foto);
-        return view('editInformasi', compact('kemajuans','proyek','foto'));
+        $pekerjaan = DB::table('jenis_pekerjaan')->where('proyek_id',$idProyek[0]->proyek_id)->get();
+        return view('editInformasi', compact('kemajuans','proyek','foto','pekerjaan'));
     }
-
     public function updateInformasi($id, Request $request){
-        
-        
-        $allId = DB::table('listPhoto')->select('listPhoto.id')->where('kemajuan_id',$id)->whereNotIn('id',$request->listId)->get();
-        $deletedId = json_decode($allId);
-        for($i=0;$i<sizeof($allId);$i++) {
-
-            DB::table('listPhoto')->where('id',$deletedId[$i]->id)->delete();
+        //dd($request->listId);
+        if ($request->listId!=null) {
+            $allId = DB::table('listPhoto')->select('listPhoto.id')->where('kemajuan_id',$id)->whereNotIn('id',$request->listId)->get();
+            //dd($allId);
+            $deletedId = json_decode($allId);
+            //dd($deletedId);
+            for($i=0;$i<sizeof($allId);$i++) {
+                DB::table('listPhoto')->where('id',$deletedId[$i]->id)->delete();
+            }
         }
         //dd($deletedId[0]->id);
         $idPelaksanaan = KemajuanProyek::select('kemajuan_proyeks.pelaksanaan_id')->where('id',$id)->get();
         $idProyek = Pelaksanaan::select('pelaksanaans.proyek_id')->whereIn('id',$idPelaksanaan)->get();
         $data = json_decode($idProyek);
         $validator = Validator::make($request->all(),[
-            'description' => 'required',
             'reportDate' => 'required',
             'tipeKemajuan' => 'required',
             'value' => 'required',
             'pelaksanaan_id' => 'required',
             'file' => 'required|image'
     	]);
-
         $kemajuans = KemajuanProyek::find($id);
+        //dd($kemajuans);
+        $kemajuans->pekerjaan_id = $request->tipepekerjaan;
         $kemajuans->description = $request->description;
         $kemajuans->reportDate = $request->reportdate;
         $kemajuans->value = $request->nilai;
         $kemajuans->tipeKemajuan = $request->tipekemajuan;
         $kemajuans->save();
-
+        //dd($kemajuans->pekerjaan_id);
+        $realValue = DB::table('kemajuan_proyeks')->select('kemajuan_proyeks.value')->where('pekerjaan_id',$kemajuans->pekerjaan_id)->get();
+        //dd(json_decode($realValue));
+        $combinedValue = 0;
+        foreach($realValue as $value) {
+            $combinedValue += $value->value;
+        }
+        DB::table('jenis_pekerjaan')->where('id',$kemajuans->pekerjaan_id)->update([
+            'workCurrentValue' => $combinedValue
+        ]);
         return redirect()->action('KemajuanProyekController@viewInfo',['id'=>$data[0]->proyek_id]);
     }
-
     public function hapusInformasi($id){
         $idPelaksanaan = KemajuanProyek::select('kemajuan_proyeks.pelaksanaan_id')->where('id',$id)->get();
         $idProyek = Pelaksanaan::select('pelaksanaans.proyek_id')->whereIn('id',$idPelaksanaan)->get();
         $data = json_decode($idProyek);
         $kemajuans = KemajuanProyek::find($id);
+        //dd($kemajuans->pekerjaan_id);
         $kemajuans->delete();
+        $realValue = DB::table('kemajuan_proyeks')->select('kemajuan_proyeks.value')->where('pekerjaan_id',$kemajuans->pekerjaan_id)->get();
+        //dd(json_decode($realValue));
+        $combinedValue = 0;
+        foreach($realValue as $value) {
+            //dd($value->value);
+            $combinedValue += $value->value;
+        }
+        //dd($combinedValue);
+        DB::table('jenis_pekerjaan')->where('id',$kemajuans->pekerjaan_id)->update([
+            'workCurrentValue' => $combinedValue
+        ]);
         return redirect()->action('KemajuanProyekController@viewInfo',['id'=>$data[0]->proyek_id]);
     }
-
     public function waktu($tanggal){
